# [Draft] Speedup dependencies installation via UV