--- a/cvat/requirements/README.md
+++ b/cvat/requirements/README.md
@@ -1,0 +1,2 @@
+These files are here for backwards compatibility only and not used during actual build.
+See ../../pyproject.toml