--- a/cvat/requirements/all.in
+++ b/cvat/requirements/all.in
@@ -1,3 +1,0 @@
--r development.in
--r production.in
--r testing.in