--- a/cvat/requirements/base.txt
+++ b/cvat/requirements/base.txt
@@ -341,6 +341,8 @@ s3transfer==0.4.2
     # via boto3
 scipy==1.14.1
     # via datumaro
+setuptools==75.3.0
+    # via cvat (pyproject.toml)
 shapely==1.7.1
     # via cvat (pyproject.toml)
 six==1.16.0