--- a/cvat/requirements/production.in
+++ b/cvat/requirements/production.in
@@ -1,7 +1,0 @@
--r base.in
-
-# To collect the coverage for REST API tests.
-# It will be enabled only if COVERAGE_PROCESS_START env is defined.
-coverage==7.2.3
-
-uvicorn[standard]==0.22.0