--- a/cvat/requirements/production.txt
+++ b/cvat/requirements/production.txt
@@ -3,20 +3,10 @@
 
 anyio==4.6.2.post1
     # via watchfiles
-async-timeout==5.0.0
-    # via redis
-azure-core==1.32.0
-    # via
-    #   azure-storage-blob
-    #   msrest
 coverage==7.2.3
     # via cvat (pyproject.toml)
 exceptiongroup==1.2.2
     # via anyio
-google-api-core==2.22.0
-    # via
-    #   google-cloud-core
-    #   google-cloud-storage
 h11==0.14.0
     # via uvicorn
 httptools==0.6.4