--- a/cvat/requirements/testing.in
+++ b/cvat/requirements/testing.in
@@ -1,4 +1,0 @@
--r development.in
-
-coverage==7.2.3
-fakeredis[lua]==2.10.3