--- a/dev/format_python_code.sh
+++ b/dev/format_python_code.sh
@@ -20,6 +20,7 @@ fi
 for paths in \
     "cvat-sdk" \
     "cvat-cli" \
+    "dev" \
     "tests/python/" \
     "cvat/apps/quality_control" \
     "cvat/apps/analytics_report" \