--- a/pyproject.toml
+++ b/pyproject.toml
@@ -56,6 +56,7 @@ dependencies = [
     "Shapely==1.7.1",
     "xmlsec>=1.3.14,<2",
     "tqdm>=4.58.0",
+    "setuptools>=75.2.0",
 ]
 
 [project.optional-dependencies]