… cvat/requirements/all.txt
… cvat/requirements/base.in
… cvat/requirements/development.in
… uv.lock