<!-- Raise an issue to propose your change (https://github.com/cvat-ai/cvat/issues).
It helps to avoid duplication of efforts from multiple independent contributors.
Discuss your ideas with maintainers to be sure that changes will be approved and merged.
Read the [Contribution guide](https://docs.cvat.ai/docs/contributing/). -->

<!-- Provide a general summary of your changes in the Title above -->

## Motivation and context
Currently, CVAT server image takes 434 seconds to build without cache on my local machine and 325 seconds after any change to requirements/base.txt. 

This extended build time is inefficient, especially when changes in dependencies or minor adjustments require almost a full rebuild.

Long builds waste developer time, delay testing cycles, and add unnecessary friction to iterative development. We need a way to optimize build speed and reduce the image size.


Following is a complete build log from my machine for `Dockerfile` from `develop` compared to `Dockerfile` from this branch:
### baseline (building images using Dockerfile from `develop`)
Resulting image size for production: `1.37GB`

<details>
<summary>build <code>--no-cache</code> — 434.4s</summary>

```shell
docker build . -t cvat_pip_no_cache --no-cache
[+] Building 434.4s (46/46) FINISHED                                                                                                                         docker:orbstack
 => [internal] load build definition from Dockerfile                                                                                                                    0.0s
 => => transferring dockerfile: 7.20kB                                                                                                                                  0.0s
 => [internal] load metadata for docker.io/library/ubuntu:22.04                                                                                                         1.4s
 => [internal] load metadata for docker.io/library/golang:1.23.0                                                                                                        1.5s
 => [internal] load .dockerignore                                                                                                                                       0.0s
 => => transferring context: 198B                                                                                                                                       0.0s
 => CACHED [build-smokescreen 1/3] FROM docker.io/library/golang:1.23.0@sha256:acfb46be39840f8c2a6b9efdd673c6627011200c73bab4e6d18b8b9ab4641c46                         0.0s
 => [internal] load build context                                                                                                                                       0.2s
 => => transferring context: 79.69kB                                                                                                                                    0.1s
 => CACHED [build-image-base 1/3] FROM docker.io/library/ubuntu:22.04@sha256:0e5e4a57c2499249aafc3b40fcd541e9a456aab7296681a3994d631587203f97                           0.0s
 => [build-smokescreen 2/3] RUN git clone --filter=blob:none --no-checkout https://github.com/stripe/smokescreen.git                                                    2.5s
 => [stage-4  2/22] RUN apt-get update &&     DEBIAN_FRONTEND=noninteractive apt-get --no-install-recommends install -yq         bzip2         ca-certificates         61.6s
 => [build-image-base 2/3] RUN apt-get update &&     DEBIAN_FRONTEND=noninteractive apt-get --no-install-recommends install -yq         curl         g++         gcc   62.8s
 => [build-smokescreen 3/3] RUN cd smokescreen && git checkout eb1ac09 && go build -o /tmp/smokescreen                                                                 12.2s
 => [stage-4  3/22] COPY --from=build-smokescreen /tmp/smokescreen /usr/local/bin/smokescreen                                                                           0.3s
 => [stage-4  4/22] RUN adduser --shell /bin/bash --disabled-password --gecos "" django                                                                                 0.5s
 => [stage-4  5/22] RUN if [ "no" = "yes" ]; then         apt-get update &&         apt-get --no-install-recommends install -yq             clamav             libclam  0.3s
 => [build-image-base 3/3] RUN --mount=type=cache,target=/root/.cache/pip/http     python3 -m pip install -U pip==24.0                                                  2.1s
 => [stage-4  6/22] RUN python3 -m venv /opt/venv                                                                                                                       2.1s
 => [build-image-av 1/9] WORKDIR /tmp/openh264                                                                                                                          0.1s
 => [build-image 1/4] COPY cvat/requirements/ /tmp/cvat/requirements/                                                                                                   0.1s
 => [stage-4  7/22] RUN python -m pip install --upgrade setuptools                                                                                                      1.9s
 => [build-image-av 2/9] RUN curl -sL https://github.com/cisco/openh264/archive/v2.1.1.tar.gz --output - |     tar -zx --strip-components=1 &&     make -j5 && make i  12.5s
 => [build-image 2/4] COPY utils/dataset_manifest/requirements.txt /tmp/utils/dataset_manifest/requirements.txt                                                         0.2s
 => [build-image 3/4] RUN sed -i '/^av==/d' /tmp/utils/dataset_manifest/requirements.txt                                                                                0.2s
 => [build-image 4/4] RUN --mount=type=cache,target=/root/.cache/pip/http-v2     DATUMARO_HEADLESS=1 python3 -m pip wheel --no-deps --no-binary lxml,xmlsec     -r /  344.5s
 => [stage-4  8/22] RUN python -m pip install -U pip==24.0                                                                                                              2.6s
 => [build-image-av 3/9] WORKDIR /tmp/ffmpeg                                                                                                                            0.1s
 => [build-image-av 4/9] RUN curl -sL https://ffmpeg.org/releases/ffmpeg-4.3.1.tar.gz --output - |     tar -zx --strip-components=1 &&     ./configure --disable-nonf  93.2s
 => [build-image-av 5/9] COPY utils/dataset_manifest/requirements.txt /tmp/utils/dataset_manifest/requirements.txt                                                      0.1s
 => [build-image-av 6/9] RUN grep -q '^av==' /tmp/utils/dataset_manifest/requirements.txt                                                                               0.1s
 => [build-image-av 7/9] RUN sed -i '/^av==/!d' /tmp/utils/dataset_manifest/requirements.txt                                                                            0.1s
 => [build-image-av 8/9] RUN pip install setuptools wheel 'cython<3'                                                                                                    1.7s
 => [build-image-av 9/9] RUN --mount=type=cache,target=/root/.cache/pip/http-v2     python3 -m pip wheel --no-binary=av --no-build-isolation     -r /tmp/utils/datase  45.1s
 => [stage-4  9/22] RUN --mount=type=bind,from=build-image,source=/tmp/wheelhouse,target=/mnt/wheelhouse     --mount=type=bind,from=build-image-av,source=/tmp/wheelh  17.6s
 => [stage-4 10/22] COPY --from=build-image-av /opt/ffmpeg/lib /usr/lib                                                                                                 0.1s
 => [stage-4 11/22] RUN if [ "${CVAT_DEBUG_ENABLED}" = 'yes' ]; then         python3 -m pip install --no-cache-dir debugpy;     fi                                      0.1s
 => [stage-4 12/22] RUN python -m pip uninstall -y pip                                                                                                                  0.4s
 => [stage-4 13/22] COPY cvat/nginx.conf /etc/nginx/nginx.conf                                                                                                          0.0s
 => [stage-4 14/22] COPY --chown=django components /tmp/components                                                                                                      0.0s
 => [stage-4 15/22] COPY --chown=django supervisord/ /home/django/supervisord                                                                                           0.0s
 => [stage-4 16/22] COPY --chown=django wait-for-it.sh manage.py backend_entrypoint.sh wait_for_deps.sh /home/django/                                                   0.1s
 => [stage-4 17/22] COPY --chown=django utils/ /home/django/utils                                                                                                       0.1s
 => [stage-4 18/22] COPY --chown=django cvat/ /home/django/cvat                                                                                                         0.4s
 => [stage-4 19/22] COPY --chown=django rqscheduler.py /home/django                                                                                                     0.1s
 => [stage-4 20/22] RUN if [ "${COVERAGE_PROCESS_START}" ]; then         echo "import coverage; coverage.process_startup()" > /opt/venv/lib/python3.10/site-packages/c  0.1s
 => [stage-4 21/22] WORKDIR /home/django                                                                                                                                0.1s
 => [stage-4 22/22] RUN mkdir -p data share keys logs /tmp/supervisord static                                                                                           0.2s
 => exporting to image                                                                                                                                                  2.4s
 => => exporting layers                                                                                                                                                 2.3s
 => => writing image sha256:ba06ec7ca5662cdcfe516c3f0f5a6a4aa480cb24bb982995355bc0e581c40fe1                                                                            0.0s
 => => naming to docker.io/library/cvat_pip_no_cache                                         
```
</details>


<details>
<summary>build after changing <code>requirements.txt</code> — 325.4s</summary>

```shell
.venv ❯ docker build .
[+] Building 325.4s (46/46) FINISHED                                                                                                                         docker:orbstack
 => [internal] load build definition from Dockerfile                                                                                                                    0.0s
 => => transferring dockerfile: 7.20kB                                                                                                                                  0.0s
 => [internal] load metadata for docker.io/library/golang:1.23.0                                                                                                        1.4s
 => [internal] load metadata for docker.io/library/ubuntu:22.04                                                                                                         1.4s
 => [internal] load .dockerignore                                                                                                                                       0.0s
 => => transferring context: 198B                                                                                                                                       0.0s
 => [build-smokescreen 1/3] FROM docker.io/library/golang:1.23.0@sha256:acfb46be39840f8c2a6b9efdd673c6627011200c73bab4e6d18b8b9ab4641c46                                0.0s
 => [internal] load build context                                                                                                                                       0.1s
 => => transferring context: 75.31kB                                                                                                                                    0.0s
 => [build-image-base 1/3] FROM docker.io/library/ubuntu:22.04@sha256:0e5e4a57c2499249aafc3b40fcd541e9a456aab7296681a3994d631587203f97                                  0.0s
 => CACHED [build-image-base 2/3] RUN apt-get update &&     DEBIAN_FRONTEND=noninteractive apt-get --no-install-recommends install -yq         curl         g++         0.0s
 => CACHED [build-image-base 3/3] RUN --mount=type=cache,target=/root/.cache/pip/http     python3 -m pip install -U pip==24.0                                           0.0s
 => [build-image 1/4] COPY cvat/requirements/ /tmp/cvat/requirements/                                                                                                   0.2s
 => [build-image 2/4] COPY utils/dataset_manifest/requirements.txt /tmp/utils/dataset_manifest/requirements.txt                                                         0.1s
 => [build-image 3/4] RUN sed -i '/^av==/d' /tmp/utils/dataset_manifest/requirements.txt                                                                                0.3s
 => [build-image 4/4] RUN --mount=type=cache,target=/root/.cache/pip/http-v2     DATUMARO_HEADLESS=1 python3 -m pip wheel --no-deps --no-binary lxml,xmlsec     -r /  295.0s
 => CACHED [build-image-av 1/9] WORKDIR /tmp/openh264                                                                                                                   0.0s
 => CACHED [build-image-av 2/9] RUN curl -sL https://github.com/cisco/openh264/archive/v2.1.1.tar.gz --output - |     tar -zx --strip-components=1 &&     make -j5 &&   0.0s
 => CACHED [build-image-av 3/9] WORKDIR /tmp/ffmpeg                                                                                                                     0.0s
 => CACHED [build-image-av 4/9] RUN curl -sL https://ffmpeg.org/releases/ffmpeg-4.3.1.tar.gz --output - |     tar -zx --strip-components=1 &&     ./configure --disabl  0.0s
 => CACHED [build-image-av 5/9] COPY utils/dataset_manifest/requirements.txt /tmp/utils/dataset_manifest/requirements.txt                                               0.0s
 => CACHED [build-image-av 6/9] RUN grep -q '^av==' /tmp/utils/dataset_manifest/requirements.txt                                                                        0.0s
 => CACHED [build-image-av 7/9] RUN sed -i '/^av==/!d' /tmp/utils/dataset_manifest/requirements.txt                                                                     0.0s
 => CACHED [build-image-av 8/9] RUN pip install setuptools wheel 'cython<3'                                                                                             0.0s
 => CACHED [build-image-av 9/9] RUN --mount=type=cache,target=/root/.cache/pip/http-v2     python3 -m pip wheel --no-binary=av --no-build-isolation     -r /tmp/utils/  0.0s
 => CACHED [stage-4  2/22] RUN apt-get update &&     DEBIAN_FRONTEND=noninteractive apt-get --no-install-recommends install -yq         bzip2         ca-certificates   0.0s
 => CACHED [build-smokescreen 2/3] RUN git clone --filter=blob:none --no-checkout https://github.com/stripe/smokescreen.git                                             0.0s
 => CACHED [build-smokescreen 3/3] RUN cd smokescreen && git checkout eb1ac09 && go build -o /tmp/smokescreen                                                           0.0s
 => CACHED [stage-4  3/22] COPY --from=build-smokescreen /tmp/smokescreen /usr/local/bin/smokescreen                                                                    0.0s
 => CACHED [stage-4  4/22] RUN adduser --shell /bin/bash --disabled-password --gecos "" django                                                                          0.0s
 => CACHED [stage-4  5/22] RUN if [ "no" = "yes" ]; then         apt-get update &&         apt-get --no-install-recommends install -yq             clamav               0.0s
 => CACHED [stage-4  6/22] RUN python3 -m venv /opt/venv                                                                                                                0.0s
 => CACHED [stage-4  7/22] RUN python -m pip install --upgrade setuptools                                                                                               0.0s
 => CACHED [stage-4  8/22] RUN python -m pip install -U pip==24.0                                                                                                       0.0s
 => [stage-4  9/22] RUN --mount=type=bind,from=build-image,source=/tmp/wheelhouse,target=/mnt/wheelhouse     --mount=type=bind,from=build-image-av,source=/tmp/wheelh  22.1s
 => [stage-4 10/22] COPY --from=build-image-av /opt/ffmpeg/lib /usr/lib                                                                                                 0.1s
 => [stage-4 11/22] RUN if [ "${CVAT_DEBUG_ENABLED}" = 'yes' ]; then         python3 -m pip install --no-cache-dir debugpy;     fi                                      0.2s
 => [stage-4 12/22] RUN python -m pip uninstall -y pip                                                                                                                  0.8s
 => [stage-4 13/22] COPY cvat/nginx.conf /etc/nginx/nginx.conf                                                                                                          0.1s
 => [stage-4 14/22] COPY --chown=django components /tmp/components                                                                                                      0.1s
 => [stage-4 15/22] COPY --chown=django supervisord/ /home/django/supervisord                                                                                           0.1s
 => [stage-4 16/22] COPY --chown=django wait-for-it.sh manage.py backend_entrypoint.sh wait_for_deps.sh /home/django/                                                   0.1s
 => [stage-4 17/22] COPY --chown=django utils/ /home/django/utils                                                                                                       0.1s
 => [stage-4 18/22] COPY --chown=django cvat/ /home/django/cvat                                                                                                         0.6s
 => [stage-4 19/22] COPY --chown=django rqscheduler.py /home/django                                                                                                     0.1s
 => [stage-4 20/22] RUN if [ "${COVERAGE_PROCESS_START}" ]; then         echo "import coverage; coverage.process_startup()" > /opt/venv/lib/python3.10/site-packages/c  0.1s
 => [stage-4 21/22] WORKDIR /home/django                                                                                                                                0.1s
 => [stage-4 22/22] RUN mkdir -p data share keys logs /tmp/supervisord static                                                                                           0.1s
 => exporting to image                                                                                                                                                  2.9s
 => => exporting layers                                                                                                                                                 2.8s
 => => writing image sha256:d58a86916699e3dcb80e817d6efc6886af3a747f1c6eb854e83252206ea3dc3a      
```
</details>


<details>
<summary>build with layer cache — 1.8s</summary>

```shell
.venv ❯ docker build .
[+] Building 1.8s (46/46) FINISHED                                                                                                                           docker:orbstack
…